pyproject.toml
[tool.poetry]
name = "quantbt"
version = "0.7.0"
description = ""
authors = ["Alpha <piotr@piotryordanov.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numba = "^0.57.0"
numpy = "^1.24.3"
pandas = "^2.0.1"
ta-lib = "^0.4.26"
matplotlib = "^3.7.1"
pyarrow = "^12.0.0"
mplfinance = "^0.12.9b7"
tqdm = "^4.65.0"
prettytable = "^3.8.0"

python-binance = {version = "^1.0.17", optional = true}
quantstats = {version = "^0.0.59", optional = true}
seaborn = {version = "^0.12.2", extras = ["stats"], optional = true}
ipython = {version = "^8.14.0", optional = true}
pandas-ta = {version = "^0.3.14b0", optional = true}
pytest = "^7.4.0"
pytest-watcher = "^0.3.4"
pyright = "^1.1.320"
importlib-resources = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
dev_tools = ["python-binance", "quantstats", "seaborn", "ipython", "pandas-ta"]

[tool.pytest-watcher]
now = true
runner = "pytest"
runner_args = []
patterns = ["*.py"]
ignore_patterns = []

[tool.pytest.ini_options]
pythonpath = "quantbt"
addopts = "-s"
testpaths = [
    "tests",
]
